+++ R/get_prism_monthlys.R
@@ -69,27 +69,25 @@ get_prism_monthlys <- function(type, years = NULL, mon = NULL, keepZip = TRUE,
       uri_dates_post81,
       function(x) {
         paste(
-          service,type,x,sep="/"
+          service, type, x, sep="/"
         )
       }
     )
   }
     
-  # download_pb <- txtProgressBar(
-    # min = 0, 
-    # max = length(uris_post81) + length(uris_pre81), 
-    # style = 3
-  # )
+  download_pb <- txtProgressBar(
+    min = 0,
+    max = length(uris_post81) + length(uris_pre81),
+    style = 3
+  )
  
   counter <- 0
 
   ### Handle post 1981 data
   if(length(uris_post81) > 0){    
-  
       for(i in 1:length(uris_post81)){
-      prism_webservice(uris_post81[i],keepZip)
-      # setTxtProgressBar(download_pb, i)
-      
+        prism_webservice(uris_post81[i],keepZip)
+        setTxtProgressBar(download_pb, i)
     }
   }
     
@@ -109,7 +107,7 @@ get_prism_monthlys <- function(type, years = NULL, mon = NULL, keepZip = TRUE,
       if (!is.null(tmp)) {
         pre_files <- c(pre_files, tmp)
       }
-      # setTxtProgressBar(download_pb, counter) 
+      setTxtProgressBar(download_pb, counter) 
       counter <- counter + 1
     }
     
@@ -136,5 +134,5 @@ get_prism_monthlys <- function(type, years = NULL, mon = NULL, keepZip = TRUE,
     }
   }
 
-  # close(download_pb)
+  close(download_pb)
 }